// src/types/slabType.ts
export type SlabType = {
  id: string;
  planReference: string;
  location_x: number;
  location_y: number;
  location_z: number;
  strength: string;
  dimensions_l: number;
  dimensions_w: number;
  dimensions_h: number;
  weight: number;
  liveload: number;
  typeOfElement: string;
  rebarDiameterTop: number;
  rebarAmountTop: number;
  rebarDiameterBottom: number;
  rebarAmountBottom: number;
};